Remove HomeController
