+++ src/ClientApp/src/components/pages/Home.tsx
@@ -21,7 +21,6 @@ import DetailsIcon from "@mui/icons-material/Details";
 import nameof from "ts-nameof.macro";
 import { InspectionSheet, InspectionSheetInitialState } from "../../entities";
 import {
-  IHomeController,
   IInspectionGroupInteractor,
   IInspectionTypeInteractor,
   IInspectionSheetInteractor,
@@ -31,10 +30,26 @@ import { SheetSearchMenu } from "../SheetSearchMenu";
 import { SheetDeleteConfirmationDialog } from "../dialog/SheetDeleteConfirmationDialog";
 import { useDIContext } from "../../container";
 
+const exportInspectionSheet = (exportUrl: string, fileName: string): void => {
+  fetch(exportUrl)
+    .then((response) => response.blob())
+    .then((blob) => {
+      const url = URL.createObjectURL(blob);
+      const a = document.createElement("a");
+      document.body.appendChild(a);
+      a.download = fileName;
+      a.href = url;
+      a.click();
+      a.remove();
+      setTimeout(() => {
+        URL.revokeObjectURL(url);
+      }, 1e4);
+    })
+    .catch(console.error);
+};
+
 export const Home: FC = (): JSX.Element => {
   const inject = useDIContext();
-  const controller: IHomeController = inject(nameof<IHomeController>());
-
   const groupUseCase: IInspectionGroupInteractor = inject(
     nameof<IInspectionGroupInteractor>()
   );
@@ -59,8 +74,11 @@ export const Home: FC = (): JSX.Element => {
   const [rowsPerPage, setRowsPerPage] = useState(10);
 
   useEffect(() => {
-    controller
-      .fetchDisplayData()
+    Promise.all([
+      typeUseCase.fetchInspectionTypes(),
+      groupUseCase.fetchInspectionGroups(),
+      sheetUseCase.fetchAllInspectionSheets(),
+    ])
       .then(() => {
         setLoading(false);
       })
@@ -85,11 +103,13 @@ export const Home: FC = (): JSX.Element => {
    */
   const handleSearch = () => {
     const { inspectionGroup, inspectionType, sheetName } = searchOption;
-    controller.searchInspectionSheet(
-      inspectionGroup,
-      inspectionType,
-      sheetName
-    );
+    if (inspectionGroup === "" && inspectionType === "" && sheetName === "") {
+      sheetUseCase.resetSearchedInspectionSheets();
+    } else {
+      const groupIds = groupUseCase.getIds(inspectionGroup);
+      const typeIds = typeUseCase.getIds(inspectionType);
+      sheetUseCase.searchInspectionSheet(groupIds, typeIds, sheetName);
+    }
     setPage(0);
   };
 
@@ -102,7 +122,7 @@ export const Home: FC = (): JSX.Element => {
       inspectionGroup: "",
       inspectionType: "",
     });
-    controller.searchInspectionSheet("", "", "");
+    sheetUseCase.resetSearchedInspectionSheets();
     setPage(0);
   };
 
@@ -113,7 +133,9 @@ export const Home: FC = (): JSX.Element => {
 
   const handleDelete = () => {
     setOpen(false);
-    controller.removeInspectionSheet(targetSheet.sheetId).catch(console.error);
+    sheetUseCase
+      .removeInspectionSheet(targetSheet.sheetId)
+      .catch(console.error);
   };
 
   /**
@@ -175,16 +197,22 @@ export const Home: FC = (): JSX.Element => {
                     aria-label="outlined button group"
                   >
                     <Button
-                      onClick={() =>
-                        controller.exportExcelInspectionSheet(sheet)
-                      }
+                      onClick={() => {
+                        exportInspectionSheet(
+                          `v1/excel-inspection-sheets/${sheet.sheetId}`,
+                          `${sheet.sheetName}.xlsx`
+                        );
+                      }}
                     >
                       Excel
                     </Button>
                     <Button
-                      onClick={() =>
-                        controller.exportJsonInspectionSheet(sheet)
-                      }
+                      onClick={() => {
+                        exportInspectionSheet(
+                          `v1/json-inspection-sheets/${sheet.sheetId}`,
+                          `${sheet.sheetName}.json`
+                        );
+                      }}
                     >
                       JSON
                     </Button>
