+++ src/ClientApp/src/controllers/index.ts
@@ -1,2 +1,1 @@
-export * from "./HomeController";
 export * from "./InspectionSheetController";
