+++ src/ClientApp/src/interfaces/controller/index.ts
@@ -1,2 +1,1 @@
-export * from "./IHomeController";
 export * from "./IInspectionSheetController";
